requirements.txt
# Web框架
Flask==2.1.0
# 数据库
mysqlclient==2.1.0  # MySQL数据库连接器
SQLAlchemy==1.4.0  # SQL工具和ORM
# API请求
requests==2.26.0  # 用于调用DeepSeek v2.5 API
# 文件处理
python-docx==0.8.11  # 处理.docx文件
# Excel文件导出
openpyxl==3.0.9  # 处理Excel文件
# 安全性
python-dotenv==0.19.0  # 用于管理环境变量,如API密钥

